Getting rid of the old fdreadouttypes header
