+++ include/fdreadoutlibs/common.hpp
@@ -1,0 +1,62 @@
+/**
+ * @file FDReadoutTypes.hpp Common types in udaq-readout
+ *
+ * This is part of the DUNE DAQ , copyright 2020.
+ * Licensing/copyright details are in the COPYING file that you should have
+ * received with this code.
+ */
+#ifndef FDREADOUTLIBS_INCLUDE_FDREADOUTLIBS_FDREADOUTTYPES_HPP_
+#define FDREADOUTLIBS_INCLUDE_FDREADOUTLIBS_FDREADOUTTYPES_HPP_
+
+//#include "iomanager/Sender.hpp"
+//#include "iomanager/Receiver.hpp"
+
+#include "daqdataformats/FragmentHeader.hpp"
+#include "daqdataformats/SourceID.hpp"
+#include "detdataformats/daphne/DAPHNEFrame.hpp"
+#include "detdataformats/ssp/SSPTypes.hpp"
+#include "detdataformats/wib/WIBFrame.hpp"
+#include "detdataformats/wib2/WIB2Frame.hpp"
+#include "detdataformats/tde/TDE16Frame.hpp"
+#include "detdataformats/fwtp/RawTp.hpp"
+#include "triggeralgs/TriggerPrimitive.hpp"
+
+#include <cstdint> // uint_t types
+#include <memory>  // unique_ptr
+#include <vector>
+#include <cstring> // memcpy
+#include <tuple> // tie
+
+namespace dunedaq {
+namespace fdreadoutlibs {
+namespace types {
+
+
+/**
+ * @brief Convencience wrapper to take ownership over char pointers with
+ * corresponding allocated memory size.
+ * */
+struct VariableSizePayloadWrapper
+{
+  VariableSizePayloadWrapper() {}
+  VariableSizePayloadWrapper(size_t size, char* data)
+    : size(size)
+    , data(data)
+  {}
+
+  size_t size = 0;
+  std::unique_ptr<char> data = nullptr;
+};
+
+
+enum CollectionOrInduction {
+  kCollection,
+  kInduction
+};
+
+
+} // namespace types
+} // namespace fdreadoutlibs
+} // namespace dunedaq
+
+#endif // FDREADOUTLIBS_INCLUDE_FDREADOUTLIBS_FDREADOUTTYPES_HPP_
